cut memory footprint during vector creation

+++ geoblend/__init__.py
@@ -19,7 +19,7 @@ def blend(source, reference, mask, solver, gradient_multiplier=1.0):
 
     indices = np.nonzero(mask)
 
-    vector = create_vector(source.astype(np.float64), reference.astype(np.float64), mask, multiplier=gradient_multiplier)
+    vector = create_vector(source, reference, mask, multiplier=gradient_multiplier)
 
     x0 = source[indices].astype('float64')
     pixels = np.round(solver.solve(b=vector, x0=x0, tol=1e-03, accel='cg'))
